requirements-dev.txt
# Include all production requirements
-r requirements.txt

# --- Code formatting & linting ---
black==24.3.0
flake8==6.1.0
isort==5.13.2

# --- Testing ---
pytest==8.2.0
pytest-flask==1.3.0

# --- Type checking (optional) ---
mypy==1.10.0

# --- Coverage (optional) ---
coverage==7.5.1
